requirements.txt
asgiref==3.6.0
Django==4.1.5
django-rest==0.8.7
djangorestframework==3.14.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
